Ported to Python 3, started fixes from last commit breakage.
Ported to python 3 for more portability, I plan to dump all of
Nouvelle and a bunch of other needless garbage currently idle in
the system. I would also like to do a nice facelift. Pretty much
every part of CIA needs to be gone through and face-lifted.
Much of the core is a mess and hard to comprehend, this should
get better with time once I am able to redo all of stats and much
of the core parts of how the RPC servers work.

This program does not run (in fact, I am sure anything that Twisted
runs is guaranteed to be broken) and the entirety of Django is way
messed up. I need to redo most of the Django code to make it more
modern and readable. This will come in due time, my main goal for
right now is to get this runnable for other users/developers.