Fixed lots of critical core errors
Fixed a large amount of the core with basic fixes to make things work
again. Most of django works, however I suspect that I'm going to need
to dump most of the authentication framework as it's really jank.

Templates now use more of Django's static files loader and url resolvers
which will make them more flexible in the future.

A large amount of the errors were related to Database stuff on the
Twisted side of things. I've eliminated DatabaseQuote as it was
a dangerous temporary fix, More database fixes are soon to come.
(Ideally I'd like Django to manage the entirety of the databases
for migration reasons with it's ORM)

Rewrote bootstrap.sh and nginx config because they were outdated
for the existing file tree.

Django no longer has runfcgi, therefore we have to use something like
uwsgi/gunicorn which I prefer over the command since now systemd
can often launch a uwsgi instance and manage it that way as a system
service instead of some random file being run.

other fixes I'm sure I'm missing are in here. A large number of erros
was from render_to_response being removed or because python strings
are unicode by default now so encoding/decoding is annoying.