Fix stats and Django not migrating SQL models
Fixed stats (Twisted based). This is mostly so I can see how it works
and begin the port to Django. Twisted's web APIs have been the worst
to migrate as they've changed drastically over the years and they're
not the easiest to debug when something goes wrong. I imagine the same
issues that plague the twisted web tac files will plague the rpc servers.

Fixed Django not migrating SQL models. I apparently needed to make a
migrations folder for everything that interacted with the database.

Worked on some more account stuff, I think registration works now but I've
not fully tested it, much of it's code was needlessly complicated.

Did some stylization stuff, this will have some more work done in the
future but for now it fixes some of the "brokenness" feeling to the site.

Migrated *all* the templates to the static files loader, we can now move
the resources out of /media/ and into /static/ so that /media/ can be
used for user-uploaded files and /static/ can be a useful cached url in
nginx.

Fixed a crash in Nouvelle. I'll be glad when I can ditch the entire
twisted web and Nouvelle mess from this codebase.